Create Module 1 : find_season
The find_season function has been implemented as Module 1. It requires two parameters: month (current month) and country (name of the nation). The function stores the seasonal variations for several nations in a dictionary called seasons.

The function looks up the specified nation in the seasons dictionary first. If yes, it uses integer division and modulo operations to determine the season index depending on the month. The appropriate season for the nation and index is then retrieved from the seasons dictionary. The test case in this code has the supplied data, with the nation set to "France" and the month set to 6. For France in June, the method returns "Summer" as the appropriate season.